Adding a time out parameter to Tab / Worker
Modifications to the Worker class and Tab class to add a timeout value that can be set per device/worker. This was done to support the changes in the PulseBlaster Labscript device files to fix a random error in which every now and then the software would hang indefinitely and cause the Pulseblaster device in BLACS to become unresponsive for days on end.