Windows Device Reset Integration
This is an attempt to integrate the devcon fix for NI Bandwith errors into blacs itself. A button is added to every tab that when clicked, will ask the user for the Windows Device ID associated with the device. Once entered, if the user has installed the necessary requirements (devcon), blacs will instruct windows to remove and re-initialize said device through the device manager. Essentially unplugging the device and plugging it back in. Once finished blacs will restart the tab through normal means. If devcon cannot be located, or other requirements are not met, or the OS isn't a windows system, an error will be noted and the tab will simply reset. Blacs will also remember the entered device ID so this only has ot be done once.

To use said feature the Windows SDK https://developer.microsoft.com/en-us/windows/downloads/windows-10-sdk/ and the Windows Development Kit https://go.microsoft.com/fwlink/?linkid=2128854 must be installed.

The windows device ID can be found in the device manager. The information can be found on the details tab for the device under "Device Instance Path". The Device ID is the value between the & and second \ characters. It could also be between two \ \ characters. So, for example, if it said USB\VID_0424&PID_2514\5&1E7355BF&0&5 you would write down “PID_2514”. If it said USB\ROOT_HUB10\5&1E7355BF&0&5 you would write down “ROOT_HUB10”.